refactor: make home page mostly server-side components
